Renamed files for `JoltPinJointImpl3D`
